Add `disable_django_management_command` flag to the config.
