+++ README.md
@@ -11,7 +11,7 @@ Commands can be run sequentially by `dj` (e.g. `dj makemigrations migrate`). How
 			"name": "nice name for the command",
 			"help": "help text for the command (optional)",  
 			"execute": "shell command to run",
-			"long_running": true  // whether the process is expected to execute and exit or run forever (optional, and defaults to `false`)
+			"long_running": true  // whether the process is expected to execute and exit or run forever (optional, defaults to false)
 		},
 		{
 			"name": "m",
@@ -24,7 +24,8 @@ Commands can be run sequentially by `dj` (e.g. `dj makemigrations migrate`). How
 			"execute": "./manage.py runserver",
 			"long_running": true
 		}
-	]
+	],
+	"disable_django_management_command": false  // prevent falling back to a Django management command cannot be found (optional, defaults to false)
 }
 ```
