+++ dj/__main__.py
@@ -2,34 +2,10 @@
 import json
 from pathlib import Path
 
-import attr
 import click
-
-from . import process_runner
+from dj import objects, process_runner
 
 DJ_CONFIG_FILE_NAME = ".dj-config.json"
-
-
-@attr.s
-class Command(object):
-    """
-    Represents a command with all of the pieces that are required.
-    """
-
-    execute = attr.ib()
-    name = attr.ib()
-    help = attr.ib(default="")
-    long_running = attr.ib(default=None)
-
-
-@attr.s
-class Config(object):
-    """
-    Stores config which is basically just a list of commands.
-    """
-
-    commands = attr.ib(default=[])
-    path = attr.ib(default="")
 
 
 @click.command()
@@ -91,11 +67,14 @@ def run(command_names, config_path, list, dry_run, verbose):
                 command = _command
                 break
 
-        if not command:
+        if not command and not config.disable_django_management_command:
             django_command_name = f"python manage.py {command_name}"
-            command = Command(execute=django_command_name, name=django_command_name)
+            command = objects.Command(
+                execute=django_command_name, name=django_command_name
+            )
 
-        process_runner.run(command, dry_run)
+        if command:
+            process_runner.run(command, dry_run)
 
 
 def _get_config_path(config_filename, verbose):
@@ -119,14 +98,15 @@ def _get_config(config_filename, verbose):
     """
     Loads the config file and serializes it into a Config object.
     """
-    path = _get_config_path(config_filename, verbose)
-    config = Config(path=str(path))
+    config_file_path = _get_config_path(config_filename, verbose)
+    config = objects.Config(config_file_path=str(config_file_path))
 
-    if path:
+    if config_file_path:
         try:
-            with path.open() as dj_config_file:
+            with config_file_path.open() as dj_config_file:
                 dj_config_text = dj_config_file.read()
                 dj_config = json.loads(dj_config_text)
+                config.disable_django_management_command = dj_config.get("disable_django_management_command")
 
                 for dj_command in dj_config.get("commands", []):
                     command_name = dj_command.get("name")
@@ -144,7 +124,7 @@ def _get_config(config_filename, verbose):
 
                         continue
 
-                    command = Command(
+                    command = objects.Command(
                         name=command_name,
                         help=dj_command.get("help"),
                         execute=execute,
@@ -154,13 +134,15 @@ def _get_config(config_filename, verbose):
                     config.commands.append(command)
 
             if verbose:
-                click.secho(f"Using {path} config file")
+                click.secho(f"Using {config_file_path} config file")
         except FileNotFoundError:
             if verbose:
                 click.secho(f"A {DJ_CONFIG_FILE_NAME} could not be found.", fg="yellow")
         except json.decoder.JSONDecodeError:
             if verbose:
-                click.secho(f"{path} does not appear to be valid JSON.", fg="yellow")
+                click.secho(
+                    f"{config_file_path} does not appear to be valid JSON.", fg="yellow"
+                )
 
     return config
