+++ dj/objects.py
@@ -1,0 +1,24 @@
+import attr
+
+
+@attr.s
+class Command(object):
+    """
+    Represents a command with all of the pieces that are required.
+    """
+
+    execute = attr.ib()
+    name = attr.ib()
+    help = attr.ib(default="")
+    long_running = attr.ib(default=None)
+
+
+@attr.s
+class Config(object):
+    """
+    Stores config which is basically just a list of commands.
+    """
+
+    commands = attr.ib(default=[])
+    config_file_path = attr.ib(default="")
+    disable_django_management_command = attr.ib(default=False)
